--- client/src/components/intro/about/about.component.jsx
@@ -1,18 +1,0 @@
-import React from "react";
-
-/* import smoothScroll from "./smoothScroll" */
-import "./about.component.styles.scss";
-
-
-const About = () => {
-
-  return (
-    <div className="about-wrapper">
-      <h2>This page is Under Construction</h2>
-      <img class="page__under-img" src="./img/under.svg" alt="" />
-
-    </div>
-  );
-};
-
-export default About;
