--- client/src/components/intro/intro.styles.scss
@@ -1,7 +1,0 @@
-.intro {
-    /* grid */
-    grid-row: 1 / 11;
-    grid-column: 1;
-    /*  */
-    //@include debug-layout;
-  }
