+++ client/src/pages/offers/list.item.component.jsx
