+++ client/src/pages/offers/list.item.styles.scss
