+++ client/src/pages/offers/offers.component.jsx
