+++ client/src/pages/offers/offers.component.styles.scss
