+++ client/src/redux/filters.reducer.js
@@ -1,6 +1,7 @@
 const INITIAL_STATE = {
     input: null,
-    mood: null
+    mood: null,
+    menuItem: 'places',
 }
 
 const searchReducer = (state = INITIAL_STATE, action) => {
@@ -14,6 +15,12 @@ const searchReducer = (state = INITIAL_STATE, action) => {
                 return {
                     ...state,
                     mood: (action.payload) ? action.payload : null
+                };
+        
+        case 'UPDATE_MENU_ITEM':
+                return {
+                    ...state,
+                    menuItem: (action.payload) ? action.payload : 'places'
                 };
         default:
             return state;
