+++ client/src/scss/_app.scss
@@ -22,3 +22,4 @@ body {
   padding: 2vh;
   background-color: #181818;
 }
+
